Added a way to make subplots, and put it to use.

+++ worldline/analyze_data/analyze_worldlines.py
@@ -13,6 +13,7 @@ import matplotlib
 matplotlib.use('PDF')
 import matplotlib.pyplot as plt
 import matplotlib.cm as cm
+import matplotlib.gridspec as gridspec
 import matplotlib.patheffects as path_effects
 
 import galaxy_diver.analyze_data.generic_data as generic_data
@@ -20,6 +21,7 @@ import galaxy_diver.plot_data.ahf as plot_ahf
 import galaxy_diver.plot_data.plotting as gen_plot
 import galaxy_diver.plot_data.pu_colormaps as pu_cm
 import galaxy_diver.utils.mp_utils as mp_utils
+import galaxy_diver.utils.utilities as utilities
 
 import analyze_ptracks
 import analyze_galids
@@ -148,30 +150,47 @@ class Worldlines( object ):
   ########################################################################
 
   def plot_hist_2d( self,
-                    x_key, y_key,
-                    slices,
-                    x_range=default, y_range=default,
-                    n_bins=128,
-                    vmin=None, vmax=None,
-                    x_label=default, y_label=default,
-                    plot_halos=False,
-                    label_plot=True,
-                    label_redshift=True,
-                    out_dir=None,
-                    *args, **kwargs ):
+    x_key, y_key,
+    slices,
+    ax = default,
+    x_range = default, y_range = default,
+    n_bins = 128,
+    vmin = None, vmax = None,
+    plot_halos = False,
+    add_colorbar = True,
+    colorbar_args = default,
+    x_label = default, y_label = default,
+    add_x_label = True, add_y_label = True,
+    plot_label = default,
+    outline_plot_label = False,
+    label_galaxy_cut = True,
+    label_redshift = True,
+    label_fontsize = 24,
+    tick_param_args = default,
+    out_dir = None,
+    *args, **kwargs ):
     '''Make a 2D histogram of the data. Extra arguments are passed to get_masked_data.
     Args:
       x_key, y_key (str) : Data keys to plot.
       slices (int or tuple of slices) : How to slices the data.
+      ax (axis) : What axis to use. By default creates a figure and places the axis on it.
       x_range, y_range ( (float, float) ) : Histogram edges. If default, all data is enclosed. If list, set manually.
         If float, is +- x_range*length scale at that snapshot.
       n_bins (int) : Number of bins in the histogram.
       vmin, vmax (float) : Limits for the colorbar.
-      x_label, ylabel_ (str) : Axes labels.
       plot_halos (bool) : Whether or not to plot merger tree halos on top of the histogram.
         Only makes sense for when dealing with positions.
-      label_plot (bool) : If True, label with self.label.
+      add_colorbar (bool) : If True, add a colorbar to colorbar_args
+      colorbar_args (axis) : What axis to add the colorbar to. By default, is ax.
+      x_label, ylabel (str) : Axes labels.
+      add_x_label, add_y_label (bool) : Include axes labels?
+      plot_label (str or dict) : What to label the plot with. By default, uses self.label.
+        Can also pass a dict of full args.
+      outline_plot_label (bool) : If True, add an outline around the plot label.
+      label_galaxy_cut (bool) : If true, add a label that indicates how the galaxy was defined.
       label_redshift (bool) : If True, add a label indicating the redshift.
+      label_fontsize (int) : Fontsize for the labels.
+      tick_param_args (args) : Arguments to pass to ax.tick_params. By default, don't change inherent defaults.
       out_dir (str) : If given, where to save the file.
     '''
 
@@ -198,16 +217,23 @@ class Worldlines( object ):
     hist2d, x_edges, y_edges = np.histogram2d( x_data, y_data, [x_edges, y_edges] )
 
     # Plot
-    fig = plt.figure( figsize=(10,9), facecolor='white' )
-    ax = plt.gca()
+    if ax is default:
+      fig = plt.figure( figsize=(10,9), facecolor='white' )
+      ax = plt.gca()
 
     im = ax.imshow( np.log10( hist2d ).transpose(), cmap=pu_cm.magma, interpolation='nearest',\
                     extent=[x_edges[0], x_edges[-1], y_edges[0], y_edges[-1]], \
                     origin='low', aspect='auto', vmin=vmin, vmax=vmax, )
 
     # Add a colorbar
-    cbar = gen_plot.add_colorbar( ax, im, method='ax' )
-    cbar.ax.tick_params( labelsize=20 )
+    if add_colorbar:
+      if colorbar_args is default:
+        colorbar_args = ax
+        cbar = gen_plot.add_colorbar( colorbar_args, im, method='ax' )
+      else:
+        colorbar_args['color_object'] = im
+        cbar = gen_plot.add_colorbar( **colorbar_args )
+      cbar.ax.tick_params( labelsize=20 )
 
     # Halo Plot
     if plot_halos:
@@ -216,28 +242,46 @@ class Worldlines( object ):
       ahf_plotter.plot_halos_snapshot( snum, ax, hubble_param=self.ptracks.data_attrs['hubble'],
         radius_fraction=self.galids.parameters['galaxy_cut'] )
       assert self.galids.parameters['length_scale'] == 'r_scale'
-      info_label = r'$r_{ \rm cut } = ' + '{:.3g}'.format( self.galids.parameters['galaxy_cut'] ) + 'r_{ s}$'
 
-    # Labels
-    if label_plot:
-      ax.annotate( s=self.label, xy=(0.,1.0225), xycoords='axes fraction', fontsize=24,  )
+    # Plot label
+    if plot_label is default:
+      plt_label = ax.annotate( s=self.label, xy=(0.,1.0225), xycoords='axes fraction', fontsize=label_fontsize,  )
+    elif isinstance( plot_label, str ):
+      plt_label = ax.annotate( s=plot_label, xy=(0.,1.0225), xycoords='axes fraction', fontsize=label_fontsize,  )
+    elif isinstance( plot_label, dict ):
+      plt_label = ax.annotate( **plot_label )
+    else:
+      raise Exception( 'Unrecognized plot_label arguments, {}'.format( plot_label ) )
+    if outline_plot_label:
+      plt_label.set_path_effects([ path_effects.Stroke(linewidth=3, foreground='black'), path_effects.Normal() ])
+
+    # Upper right label (info label)
+    info_label = ''
+    if label_galaxy_cut:
+      info_label = r'$r_{ \rm cut } = ' + '{:.3g}'.format( self.galids.parameters['galaxy_cut'] ) + 'r_{ s}$'
     if label_redshift:
       info_label = r'$z=' + '{:.3f}'.format( self.ptracks.redshift.iloc[slices] ) + '$, '+ info_label
-    if plot_halos or label_redshift:
-      ax.annotate( s=info_label, xy=(1.,1.0225), xycoords='axes fraction', fontsize=24,
+    if label_galaxy_cut or label_redshift:
+      ax.annotate( s=info_label, xy=(1.,1.0225), xycoords='axes fraction', fontsize=label_fontsize,
         ha='right' )
 
-    # Add labels
-    if x_label is default:
-      x_label = x_key
-    if y_label is default:
-      y_label = y_key
-    ax.set_xlabel( x_label, fontsize=24 )
-    ax.set_ylabel( y_label, fontsize=24 )
+    # Add axis labels
+    if add_x_label:
+      if x_label is default:
+        x_label = x_key
+      ax.set_xlabel( x_label, fontsize=label_fontsize )
+    if add_y_label:
+      if y_label is default:
+        y_label = y_key
+      ax.set_ylabel( y_label, fontsize=label_fontsize )
 
     # Limits
     ax.set_xlim( x_range )
     ax.set_ylim( y_range )
+
+    # Set tick parameters
+    if tick_param_args is not default:
+      ax.tick_params( **tick_param_args )
 
     # Save the file
     if out_dir is not None:
@@ -248,14 +292,120 @@ class Worldlines( object ):
 
   ########################################################################
 
+  def panel_plot( self,
+    panel_plotting_method_str,
+    defaults,
+    variations,
+    plot_label = default,
+    outline_plot_label = False,
+    label_galaxy_cut = True,
+    label_redshift = True,
+    label_fontsize = 24,
+    subplot_label_args = { 'xy' : (0.075, 0.88), 'xycoords' : 'axes fraction', 'fontsize' : 18, 'color' : 'w',  },
+    subplot_spacing_args = { 'hspace' : 0.0001, 'wspace' : 0.0001, },
+    ):
+    '''
+    Make a multi panel plot of the type of your choosing.
+    Note: Currently only compatible with a four panel plot.
+
+    Args:
+      panel_plotting_method_str (str) : What type of plot to make.
+      defaults (dict) : Default arguments to pass to panel_plotting_method.
+      variations (dict of dicts) : Differences in plotting arguments per subplot.
+      plot_label (str or dict) : What to label the plot with. By default, uses self.label.
+        Can also pass a dict of full args.
+      outline_plot_label (bool) : If True, add an outline around the plot label.
+      label_galaxy_cut (bool) : If true, add a label that indicates how the galaxy was defined.
+      label_redshift (bool) : If True, add a label indicating the redshift.
+      label_fontsize (int) : Fontsize for the labels.
+      subplot_label_args (dict) : Label arguments to pass to each subplot for the label for the subplot.
+        The actual label string itself corresponds to the keys in variations.
+      subplot_spacing_args (dict) : How to space the subplots.
+    '''
+
+    fig = plt.figure( figsize=(10,9), facecolor='white', )
+    ax = plt.gca()
+
+    fig.subplots_adjust( **subplot_spacing_args )
+
+    plotting_kwargs = utilities.dict_from_defaults_and_variations( defaults, variations )
+
+    # Setup axes
+    gs = gridspec.GridSpec(2, 2)
+    axs = []
+    axs.append( plt.subplot( gs[0,0] ) )
+    axs.append( plt.subplot( gs[0,1] ) )
+    axs.append( plt.subplot( gs[1,0] ) )
+    axs.append( plt.subplot( gs[1,1] ) )
+
+    # Setup arguments further
+    for i, key in enumerate( plotting_kwargs.keys() ):
+      ax_kwargs = plotting_kwargs[key]
+
+      ax_kwargs['ax'] = axs[i]
+
+      # Subplot label args
+      this_subplot_label_args = subplot_label_args.copy()
+      this_subplot_label_args['s'] = key
+      ax_kwargs['plot_label'] = this_subplot_label_args
+
+      if ax_kwargs['add_colorbar']:
+        ax_kwargs['colorbar_args'] = { 'fig_or_ax' : fig, 'ax_location' : [0.9, 0.125, 0.03, 0.775 ],  }
+
+      # Clean up interior axes
+      ax_tick_parm_args = ax_kwargs['tick_param_args'].copy()
+      if i == 0:
+        ax_kwargs['add_x_label'] = False
+        ax_tick_parm_args['labelbottom'] = False
+      if i == 1:
+        ax_kwargs['add_x_label'] = False
+        ax_tick_parm_args['labelbottom'] = False
+        ax_kwargs['add_y_label'] = False
+        ax_tick_parm_args['labelleft'] = False
+      elif i == 3:
+        ax_kwargs['add_y_label'] = False
+        ax_tick_parm_args['labelleft'] = False
+      ax_kwargs['tick_param_args'] = ax_tick_parm_args
+
+    # Actual panel plots
+    panel_plotting_method = getattr( self, panel_plotting_method_str )
+    for key in plotting_kwargs.keys():
+      panel_plotting_method( **plotting_kwargs[key] )
+
+    # Main axes labels
+    # Plot label
+    if plot_label is default:
+      plt_label = axs[0].annotate( s=self.label, xy=(0.,1.0225), xycoords='axes fraction', fontsize=label_fontsize,  )
+    elif isinstance( plot_label, str ):
+      plt_label = axs[0].annotate( s=plot_label, xy=(0.,1.0225), xycoords='axes fraction', fontsize=label_fontsize,  )
+    elif isinstance( plot_label, dict ):
+      plt_label = axs[0].annotate( **plot_label )
+    else:
+      raise Exception( 'Unrecognized plot_label arguments, {}'.format( plot_label ) )
+    if outline_plot_label:
+      plt_label.set_path_effects([ path_effects.Stroke(linewidth=3, foreground='black'), path_effects.Normal() ])
+
+    # Upper right label (info label)
+    info_label = ''
+    if label_galaxy_cut:
+      info_label = r'$r_{ \rm cut } = ' + '{:.3g}'.format( self.galids.parameters['galaxy_cut'] ) + 'r_{ s}$'
+    if label_redshift:
+      ind = defaults['slices']
+      info_label = r'$z=' + '{:.3f}'.format( self.ptracks.redshift.iloc[ind] ) + '$, '+ info_label
+    if label_galaxy_cut or label_redshift:
+      axs[1].annotate( s=info_label, xy=(1.,1.0225), xycoords='axes fraction', fontsize=label_fontsize,
+        ha='right' )
+
+  ########################################################################
+
   def make_multiple_plots( self,
-                           plotting_method_str,
-                           iter_args_key,
-                           iter_args,
-                           n_processors=1,
-                           out_dir=None,
-                           make_movie=False,
-                           *args, **kwargs ):
+    plotting_method_str,
+    iter_args_key,
+    iter_args,
+    n_processors = 1,
+    out_dir = None,
+    make_movie = False,
+    *args, **kwargs ):
     '''Make multiple plots of a selected type. *args and **kwargs are passed to plotting_method_str.
 
     Args:
@@ -315,11 +465,11 @@ class WorldlineDataMasker( generic_data.DataMasker ):
   ########################################################################
 
   def get_masked_data( self,
-                       data_key,
-                       mask=default,
-                       classification=None,
-                       mask_after_first_acc=False,
-                       *args, **kwargs ):
+    data_key,
+    mask=default,
+    classification=None,
+    mask_after_first_acc=False,
+    *args, **kwargs ):
     '''Get masked worldline data. Extra arguments are passed to the ParentClass' get_masked_data.
 
     Args:
@@ -342,7 +492,8 @@ class WorldlineDataMasker( generic_data.DataMasker ):
 
     if classification is not None:
       cl_mask = np.invert( self.worldlines.classifications.data[classification] ) 
-      cl_mask = np.tile( cl_mask, (self.worldlines.n_snaps, 1) ).transpose()
+      if classification != 'is_wind':
+        cl_mask = np.tile( cl_mask, (self.worldlines.n_snaps, 1) ).transpose()
       used_masks.append( cl_mask )
 
     if mask_after_first_acc:
